fix shell scripts

+++ osbuild-manifests/files/apply-config.sh
@@ -12,37 +12,50 @@ argument=$(systemd-escape -u "$1")
 echo "unescaped: ${argument}"
 
 op="$(cut -d '-' -f 1 <<< "$argument")"
-config_image="$(cut -d ';' -f 2- <<< "$argument")"
+config_image="$(cut -d '-' -f 2 <<< "$argument")"
 
-manifest=$(mktemp -p .)
-files_dir=$(mktemp -p . -d)
+files_dir=$(mktemp -d)
 
-podman run -t --rm -v /tmp/:/workspace ghcr.io/oras-project/oras:latest manifest fetch \
-    --plain-http \
-    ${config_image} > ${manifest}
-
-podman run -t --rm -v /tmp/:/workspace ghcr.io/oras-project/oras:latest pull \
+podman run -t --rm -v ${files_dir}:/workspace:z ghcr.io/oras-project/oras:v1.0.0 manifest fetch \
     --plain-http \
     ${config_image} \
-    --output ${files_dir}
+    --output manifest.json
 
-number_of_layers=$(jq '.layers | length' ${manifest})
+podman run -t --rm -v ${files_dir}:/workspace:z ghcr.io/oras-project/oras:v1.0.0 pull \
+    --plain-http \
+    ${config_image} \
+    --output ./
+
+number_of_layers=$(jq '.layers | length' ${files_dir}/manifest.json)
 
 for (( n=0; n<${number_of_layers}; n++ ))
 do
-    file=$(jq --argjson index ${n} '.layers[$index].annotations."org.opencontainers.image.title"' ${bundle_config} -r)
-    path=$(jq --argjson index ${n} '.layers[$index].annotations.path' ${bundle_config} -r)
-    op=$(jq --argjson index ${n} '.layers[$index].op' ${bundle_config} -r)
+    file=$(jq --argjson index ${n} '.layers[$index].annotations."org.opencontainers.image.title"' ${files_dir}/manifest.json -r)
+    path=$(jq --argjson index ${n} '.layers[$index].annotations.path' ${files_dir}/manifest.json -r)
+    mode=$(jq --argjson index ${n} '.layers[$index].annotations.mode' ${files_dir}/manifest.json -r)
 
     if [ "$op" = "$OP_CREATE" ]; then
+        if [ -f ${path} ]; then
+            echo "error: file ${path} already exists and can not be created again."
+            exit 1
+        fi
         cp -f ${files_dir}/${file} ${path}
-    else if [ "$op" = "$OP_UPDATE" ]; then
+        chmod ${mode} ${path}
+    elif [ "$op" = "$OP_UPDATE" ]; then
+        if [ ! -f ${path} ]; then
+            echo "error: file ${path} does not exist and can not be updated."
+            exit 1
+        fi
         rm ${path}
         cp -f ${files_dir}/${file} ${path}
-    else if [ "$op" = "$OP_DELETE" ]; then
+        chmod ${mode} ${path}
+    elif [ "$op" = "$OP_DELETE" ]; then
+        if [ ! -f ${path} ]; then
+            echo "error: file ${path} does not exist and can not be deleted."
+            exit 1
+        fi
         rm ${path}
     fi
 done
 
-rm ${manifest}
 rm -rf ${files_dir}
